# app/docs/api/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "GenAI Demo - DDD 電商平台 API",
    "description" : "基於領域驅動設計（DDD）的電商平台 API\n\n## 架構特色\n- 遵循 DDD 分層架構\n- 六邊形架構（Hexagonal Architecture）\n- CQRS 和事件驅動設計\n- 豐富的領域模型\n\n## 主要功能模組\n- 訂單管理：完整的訂單生命週期\n- 支付處理：多種支付方式和狀態管理\n- 庫存管理：即時庫存追蹤和預留\n- 產品管理：產品資訊和分類\n- 促銷活動：多樣化的促銷規則\n- 定價策略：動態定價和佣金計算\n- 客戶管理：會員等級和偏好設定\n\n## 錯誤處理\n所有 API 都遵循統一的錯誤回應格式，包含錯誤代碼、訊息、時間戳記和詳細資訊。\n",
    "contact" : {
      "name" : "GenAI Demo Team",
      "url" : "https://github.com/humank/genai-demo",
      "email" : "contact@genai-demo.com"
    },
    "license" : {
      "name" : "MIT License",
      "url" : "https://opensource.org/licenses/MIT"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "http://test-api.genai-demo.com",
    "description" : "測試環境"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
...
  "paths" : {
...
    "/api/consumer/cart/{customerId}/items/{productId}" : {
      "put" : {
        "tags" : [ "購物車管理" ],
        "summary" : "更新購物車商品數量",
        "description" : "更新購物車中指定商品的數量",
        "operationId" : "updateItemQuantity",
        "parameters" : [ {
          "name" : "customerId",
          "in" : "path",
          "description" : "客戶ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "productId",
          "in" : "path",
          "description" : "商品ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateQuantityRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "string"
                  }
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "購物車管理" ],
        "summary" : "從購物車移除商品",
        "description" : "從購物車中移除指定商品",
        "operationId" : "removeItemFromCart",
        "parameters" : [ {
          "name" : "customerId",
          "in" : "path",
          "description" : "客戶ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "productId",
          "in" : "path",
          "description" : "商品ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "string"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/consumer/cart/{customerId}/promotions" : {
      "get" : {
        "tags" : [ "購物車管理" ],
        "summary" : "獲取可用促銷活動",
        "description" : "獲取購物車可用的促銷活動",
        "operationId" : "getAvailablePromotions",
        "parameters" : [ {
          "name" : "customerId",
          "in" : "path",
          "description" : "客戶ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "object"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "購物車管理" ],
        "summary" : "應用促銷活動",
        "description" : "向購物車應用促銷活動",
        "operationId" : "applyPromotion",
        "parameters" : [ {
          "name" : "customerId",
          "in" : "path",
          "description" : "客戶ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ApplyPromotionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/consumer/cart/{customerId}/items" : {
      "post" : {
        "tags" : [ "購物車管理" ],
        "summary" : "添加商品到購物車",
        "description" : "向購物車添加商品",
        "operationId" : "addItemToCart",
        "parameters" : [ {
          "name" : "customerId",
          "in" : "path",
          "description" : "客戶ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AddItemRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/consumer/cart/{customerId}/coupons" : {
      "post" : {
        "tags" : [ "購物車管理" ],
        "summary" : "應用優惠券",
        "description" : "向購物車應用優惠券",
        "operationId" : "applyCoupon",
        "parameters" : [ {
          "name" : "customerId",
          "in" : "path",
          "description" : "客戶ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ApplyCouponRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "string"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/consumer/cart/{customerId}" : {
      "get" : {
        "tags" : [ "購物車管理" ],
        "summary" : "獲取購物車",
        "description" : "根據客戶ID獲取購物車信息",
        "operationId" : "getShoppingCart",
        "parameters" : [ {
          "name" : "customerId",
          "in" : "path",
          "description" : "客戶ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ShoppingCartDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/consumer/cart/{customerId}/inventory-check" : {
      "get" : {
        "tags" : [ "購物車管理" ],
        "summary" : "檢查庫存可用性",
        "description" : "檢查購物車商品的庫存可用性",
        "operationId" : "checkInventoryAvailability",
        "parameters" : [ {
          "name" : "customerId",
          "in" : "path",
          "description" : "客戶ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "object"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/consumer/cart/{customerId}/checkout" : {
      "get" : {
        "tags" : [ "購物車管理" ],
        "summary" : "計算結帳金額",
        "description" : "計算購物車的結帳金額",
        "operationId" : "calculateCheckout",
        "parameters" : [ {
          "name" : "customerId",
          "in" : "path",
          "description" : "客戶ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "object"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/consumer/cart/{customerId}/coupons/{couponCode}" : {
      "delete" : {
        "tags" : [ "購物車管理" ],
        "summary" : "移除優惠券",
        "description" : "從購物車移除優惠券",
        "operationId" : "removeCoupon",
        "parameters" : [ {
          "name" : "customerId",
          "in" : "path",
          "description" : "客戶ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "couponCode",
          "in" : "path",
          "description" : "優惠券代碼",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/consumer/cart/{customerId}/clear" : {
      "delete" : {
        "tags" : [ "購物車管理" ],
        "summary" : "清空購物車",
        "description" : "清空購物車中的所有商品",
        "operationId" : "clearCart",
        "parameters" : [ {
          "name" : "customerId",
          "in" : "path",
          "description" : "客戶ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "string"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "UpdateQuantityRequest" : {
        "type" : "object",
        "properties" : {
          "quantity" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "ApplyPromotionRequest" : {
        "type" : "object",
        "properties" : {
          "promotionId" : {
            "type" : "string"
          }
        }
      },
      "AddItemRequest" : {
        "type" : "object",
        "properties" : {
          "productId" : {
            "type" : "string"
          },
          "quantity" : {
            "type" : "integer",
            "format" : "int32"
          },
          "unitPrice" : {
            "type" : "number",
            "format" : "double"
          }
        }
      },
      "ApplyCouponRequest" : {
        "type" : "object",
        "properties" : {
          "couponCode" : {
            "type" : "string"
          }
        }
      },
...
      "CartItemDto" : {
        "type" : "object",
        "properties" : {
          "productId" : {
            "type" : "string"
          },
          "productName" : {
            "type" : "string"
          },
          "quantity" : {
            "type" : "integer",
            "format" : "int32"
          },
          "unitPrice" : {
            "type" : "number"
          },
          "totalPrice" : {
            "type" : "number"
          }
        }
      },
      "ShoppingCartDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "customerId" : {
            "type" : "string"
          },
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CartItemDto"
            }
          },
          "totalAmount" : {
            "type" : "number"
          },
          "discountAmount" : {
            "type" : "number"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "ACTIVE", "ABANDONED", "CHECKED_OUT", "EXPIRED" ]
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      }
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "description" : "請在此處輸入 JWT Token，格式：Bearer {token}",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      },
      "apiKey" : {
        "type" : "apiKey",
        "description" : "請在此處輸入 API Key",
        "name" : "X-API-Key",
        "in" : "header"
      }
    }
  }
}